Avoid needless 64-bit args on 32-bit machines for some intrinsics
* libguile/intrinsics.h:
* libguile/intrinsics.c (string_set_x): Change to take size_t and u32 as
  args.
  (allocate_words): Change to take size_t as arg.
* libguile/vm.c (expand_apply_argument): Rename from rest_arg_length,
  and also handle the stack manipulation.
* libguile/vm-engine.c (expand-apply-argument): Update for intrinsic
  change.
  (call-scm-sz-u32): Rename from call-scm-u64-u64, as it matches its
  uses and will compile better on 32-bit systems.
* module/system/vm/assembler.scm (define-scm-sz-u32-intrinsic):
  (string-set!): Update for new instrinsic call inst.
* libguile/jit.c (compile_call_scm_sz_u32): Adapt.

+++ libguile/intrinsics.c
@@ -73,7 +73,7 @@ sub_immediate (SCM a, uint8_t b)
 }
 
 static void
-string_set_x (SCM str, uint64_t idx, uint64_t ch)
+string_set_x (SCM str, size_t idx, uint32_t ch)
 {
   str = scm_i_string_start_writing (str);
   scm_i_string_set_x (str, idx, ch);
@@ -344,7 +344,7 @@ error_wrong_number_of_values (uint32_t expected)
 }
 
 static SCM
-allocate_words (scm_thread *thread, uint64_t n)
+allocate_words (scm_thread *thread, size_t n)
 {
   return SCM_PACK_POINTER (scm_inline_gc_malloc_words (thread, n));
 }
